Agregue listado de equipos de una zona determinada

+++ src/main/java/com/patricio/springboot/app/repository/EstadisticaJugadorRepository.java
@@ -3,5 +3,5 @@ package com.patricio.springboot.app.repository;
 import com.patricio.springboot.app.entity.EstadisticaJugador;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EstadisticaJugadorRepository extends JpaRepository<EstadisticaJugador, Integer> {
+public interface EstadisticaJugadorRepository extends JpaRepository<EstadisticaJugador, Long> {
 }
